Create the manager
Initially was going to do something similar to what I do except I have
to look for a different type of portal, so using the Bukkit TravelAgent
isn't going to work. Also, since we're always using a multiplier, we
don't really need to store locations, so that file is also probably
going away. So the fun part is doing what the TravelAgent does for these
new portals...